python/requirements.txt
opencv-python-headless<4.3
runpod
pytesseract
Pillow